impl top k by
